refactor(services-dotnet): group dotnet DSL types by concern
