+++ modules/dsl-services-dotnet/src/main/kotlin/io/github/lmliam/microsmith/dsl/services/dotnet/core/DotnetScopeExtensions.kt
@@ -4,6 +4,12 @@ import io.github.lmliam.microsmith.dsl.services.core.ServiceBuilder
 import io.github.lmliam.microsmith.dsl.services.core.ServiceScope
 import io.github.lmliam.microsmith.dsl.services.core.ServicesBuilder
 import io.github.lmliam.microsmith.dsl.services.core.ServicesScope
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.service.DotnetServiceBuilder
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.service.DotnetServiceExtension
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.service.DotnetServiceScope
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.shared.DotnetSharedBuilder
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.shared.DotnetSharedExtension
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.shared.DotnetSharedScope
 
 /**
  * Start a shared .NET defaults block inside `services { ... }`.
