+++ modules/dsl-services-dotnet/src/main/kotlin/io/github/lmliam/microsmith/dsl/services/dotnet/core/service/DotnetServiceContext.kt
@@ -1,4 +1,4 @@
-package io.github.lmliam.microsmith.dsl.services.dotnet.core
+package io.github.lmliam.microsmith.dsl.services.dotnet.core.service
 
 import io.github.lmliam.microsmith.dsl.services.core.ServiceExtension
 import kotlin.reflect.KClass
