+++ modules/dsl-services-dotnet/src/main/kotlin/io/github/lmliam/microsmith/dsl/services/dotnet/core/shared/DotnetSharedContext.kt
@@ -1,8 +1,8 @@
-package io.github.lmliam.microsmith.dsl.services.dotnet.core
+package io.github.lmliam.microsmith.dsl.services.dotnet.core.shared
 
 import io.github.lmliam.microsmith.dsl.core.MicrosmithExtension
 import kotlin.reflect.KClass
 
-interface DotnetSolutionContext : DotnetSolutionScope {
+interface DotnetSharedContext : DotnetSharedScope {
     fun <T : MicrosmithExtension> put(type: KClass<T>, ext: T)
 }
