+++ modules/dsl-services-dotnet/src/main/kotlin/io/github/lmliam/microsmith/dsl/services/dotnet/core/solution/DotnetSolution.kt
@@ -1,6 +1,7 @@
-package io.github.lmliam.microsmith.dsl.services.dotnet.core
+package io.github.lmliam.microsmith.dsl.services.dotnet.core.solution
 
 import io.github.lmliam.microsmith.dsl.core.MicrosmithExtension
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.support.validateDotnetQualifiedIdentifier
 import kotlin.reflect.KClass
 
 /**
