+++ modules/dsl-services-dotnet/src/main/kotlin/io/github/lmliam/microsmith/dsl/services/dotnet/core/solution/DotnetSolutionContext.kt
@@ -1,8 +1,8 @@
-package io.github.lmliam.microsmith.dsl.services.dotnet.core
+package io.github.lmliam.microsmith.dsl.services.dotnet.core.solution
 
 import io.github.lmliam.microsmith.dsl.core.MicrosmithExtension
 import kotlin.reflect.KClass
 
-interface DotnetSharedContext : DotnetSharedScope {
+interface DotnetSolutionContext : DotnetSolutionScope {
     fun <T : MicrosmithExtension> put(type: KClass<T>, ext: T)
 }
