+++ modules/dsl-services-dotnet/src/main/kotlin/io/github/lmliam/microsmith/dsl/services/dotnet/core/solution/DotnetSolutionScope.kt
@@ -1,4 +1,4 @@
-package io.github.lmliam.microsmith.dsl.services.dotnet.core
+package io.github.lmliam.microsmith.dsl.services.dotnet.core.solution
 
 import io.github.lmliam.microsmith.dsl.core.MicrosmithDsl
