+++ modules/gen-services-dotnet/src/main/kotlin/io/github/lmliam/microsmith/gen/services/dotnet/DotnetWorkspace.kt
@@ -1,7 +1,7 @@
 package io.github.lmliam.microsmith.gen.services.dotnet
 
-import io.github.lmliam.microsmith.dsl.services.dotnet.core.DotnetSolution
 import io.github.lmliam.microsmith.dsl.services.dotnet.core.DotnetTarget
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.solution.DotnetSolution
 
 /**
  * Resolved .NET workspace state after DSL normalisation.
