+++ modules/gen-services-dotnet/src/main/kotlin/io/github/lmliam/microsmith/gen/services/dotnet/ResolvedDotnetService.kt
@@ -1,8 +1,8 @@
 package io.github.lmliam.microsmith.gen.services.dotnet
 
-import io.github.lmliam.microsmith.dsl.services.dotnet.core.DotnetModel
-import io.github.lmliam.microsmith.dsl.services.dotnet.core.DotnetSolution
 import io.github.lmliam.microsmith.dsl.services.dotnet.core.DotnetTarget
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.model.DotnetModel
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.solution.DotnetSolution
 
 /**
  * Resolved per-service .NET generation state.
